{roo code} consolidate algorithm file structure
